backend/requirements.txt
# Core Web Framework
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-multipart
aiofiles

# LangChain Ecosystem
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
langchain-groq
langchain-huggingface
langchain-qdrant

# AI, Machine Learning & Embeddings
google-generativeai
sentence-transformers
transformers
torch
scikit-learn

# Vector Database
qdrant-client
faiss-cpu # Note: You have both qdrant and faiss. Ensure you need both.

# Data Handling & Utilities
pandas
numpy
requests
beautifulsoup4
python-dotenv
loguru

# Document Loaders
python-docx
pypdf

# Prescription Analysis Dependencies
pytesseract==0.3.10
Pillow==10.0.1
PyMuPDF==1.23.8
spacy==3.7.2
opencv-python==4.8.1.78
